### 2025-09-18 8960555
:style update paper.md
