### 2025-09-20 #9005 @MaximLippeveld
Happy to review as well.
