### 2025-09-26 #9005 @adamltyson
Not quite yet @pnewstein. I will wait a few more days to see if anyone else wants to review, then I will start the review officially. At that point, the reviewers will go through the paper and code, and start suggesting changes they deem necessary to complete the review checklist.
