### 2025-09-29 #9091 @editorialbot
[REVIEW]: Napari-3D-Counter: A manual cell counter for napari
**Submitting author:** @pnewstein (<a href="http://orcid.org/0000-0003-2966-783X">Peter Newstein</a>)
**Repository:** https://github.com/pnewstein/napari-3d-counter
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.9
