### 2025-10-10 e5421c2
:fix ingress points adds all out_of_slice
instead of adding directly to the layer, it now adds to pointer
also added a test to confirm this behavior
