### 2025-10-10 bb8ffe8
:fix unpin napari version
given napari's change to EffVer I am no longer pinning the micro version
of napari
