+++ setup.cfg
@@ -35,7 +35,7 @@ install_requires =
     qtpy
     pandas
     scikit-image
-    napari==0.6.4
+    napari==0.6.*
 
 python_requires = >=3.10
 include_package_data = True
