### 2025-10-24 c67851f
:fix change tested python versions
