### 2025-10-24 6313b82
:feat updated paper to add pubs that use N3DC
