+++ paper.bib
@@ -61,3 +61,32 @@
   urldate = {2025-08-26},
   langid = {english},
 }
+
+@article{drakeFer3UniquelyExpressed2025,
+  title = {Fer3 Is Uniquely Expressed in {{NotchOFF}} Hemilineages Where It Promotes Interneuron Identity},
+  author = {Drake, Kasey C. and Lai, Sen-Lin and Doe, Chris Q.},
+  date = {2025-10-22},
+  journaltitle = {Development},
+  shortjournal = {Development},
+  pages = {dev.205118},
+  issn = {0950-1991},
+  doi = {10.1242/dev.205118},
+  url = {https://doi.org/10.1242/dev.205118},
+  urldate = {2025-10-24},
+}
+
+@article{satoVIIITubulinIdentifies2025,
+  title = {β-{{III}} Tubulin Identifies Anti-Fibrotic State of Pericytes in Pulmonary Fibrosis},
+  author = {Sato, Ryo and Imamura, Kosuke and Tsukui, Tatsuya and Yoshida, Teruhiko and Tomita, Yusuke and Fujino, Kosuke and Ikeda, Tokunori and Onizawa, Keigo and Sogo, Takumi and Combs, Christian A. and Murgai, Meera and Kopp, Jeffrey B. and Suzuki, Makoto and Sakagami, Takuro and Sheppard, Dean and Mukouyama, Yoh-Suke},
+  date = {2025-04-19},
+  journaltitle = {bioRxiv: The Preprint Server for Biology},
+  shortjournal = {bioRxiv},
+  eprint = {40568101},
+  eprinttype = {pmid},
+  pages = {2025.04.15.648984},
+  issn = {2692-8205},
+  doi = {10.1101/2025.04.15.648984},
+  langid = {english},
+  pmcid = {PMC12190747}
+}
+
