### 2025-11-02 #9091 @adamltyson
Napari runs fine on Windows, could you try installing it natively?
