### 2025-11-14 pnewstein/napari-3d-counter#15 @TimMonko
First click after changing cell type does not add a point (sometimes)
During review of https://github.com/openjournals/joss-reviews/issues/9091 I have found that using the hotkeys to change cell type results in the first click not adding a point (and while writing this issue it seems sporadic) Subsequent clicks always add a point. I _think_ it might be if I change to a new cell type and then immediately click is when its not added. If I wait after changing cell type than it seems to always add one.

I am using `uvx --with napari-3d-counter napari[pyqt6,optional]` to launch napari.
