### 2025-11-14 pnewstein/napari-3d-counter#16 @TimMonko
Points added in 3D view do not have centroids and circles placed in proper location
During review of https://github.com/openjournals/joss-reviews/issues/9091 I found that points added in 2D behavior as expected, due to being positioned within a slice (all the not red ones). However, points place in 3D (the red ones) do not have the central point matching the position of the outline circle (btw what are you doing to accomplish this, two points represented in different ways via the same click -- I know napari points aren't normally styled this way).

It's also unclear to me from the documentation if this plugin is meant to function in 3D, as point placement is hard for many reasons 😉

I am using `uvx --with napari-3d-counter napari[pyqt6,optional]` to launch napari.
