### 2025-11-14 pnewstein/napari-3d-counter#17 @TimMonko
What is the purpose of the `make launch_cell_count.py` button?
Re: https://github.com/openjournals/joss-reviews/issues/9091

Pressing the buttons adds a script to a path on my computer.
If I run that script I seem to just be presented with the same widget (I guess its a dock widget this time instead of plugin dock widget), despite there being some config in the script.

It seems this time there is no hotkeys to change types, and the cells get added with a different appearance, why?
