### 2025-11-14 pnewstein/napari-3d-counter#18 @TimMonko
Consider colorblind-friendly point colors
Re: https://github.com/openjournals/joss-reviews/issues/9091

I am red green colorblind, I am finding it challenging to easily identify the difference between types 1 and 2 because of this. Consider a more typical colorblind friendly cycle like CMYRGB. I know that this is difficult with more than 3 colors, so this is a suggestion.
