### 2025-11-14 pnewstein/napari-3d-counter#19 @TimMonko
Consider the "scale" of the points counting for broader applicability
Re: https://github.com/openjournals/joss-reviews/issues/9091

I opened the simplest sample I could think of, Cells3D, in napari. Within 4 slices of adding the cell type marker, I already can't observe the plugins unique visualization. Is this intentional? If not, how might you extend this functionality to applications outside of the sample you show in the README?
