### 2025-11-14 pnewstein/napari-3d-counter#20 @TimMonko
Consider using the native file dialog instead of the Qt styled one
Re: https://github.com/openjournals/joss-reviews/issues/9091

I genuinely only offer this as a suggestion. Indeed, this is a point of contention amongst the napari core team -- we even had a joking "war" at our recent hackathon.

However, because of the nature of how napari uses the native file dialog at present, it was quite jarring to be presented with this qt styled dialog, especially because my normal way of navigating to things like favorited folders and such was lost, and I was just stuck with my base level paths conveniently.
