### 2025-11-14 pnewstein/napari-3d-counter#21 @TimMonko
Consider using a single points layer with styling for the points
Re: https://github.com/openjournals/joss-reviews/issues/9091

One way to simplify the utilization of this plugin would be to set features of added points, rather than switching between layers. See the documentation on styling points: https://napari.org/stable/howtos/layers/points.html#changing-points-border-and-face-color
Hotkeys could still be used to change, for example, `current_border_color` and result in only a single layer needed to represent all the cell types. This would potentially be a large refactor, but also simplify your io as a result. As a user, I think this would be a much clearer path forward and more intuitive, but I understand the time commitment needed to change to such behavior.
