### 2025-11-14 f8eae4c
:fix closes #16
A warning is emited to show the user that this is action is not
supported. It also does not add the point
