### 2025-11-14 pnewstein/napari-3d-counter#22 @TimMonko
Add `Load cells` as a napari reader function
Re: Re: https://github.com/openjournals/joss-reviews/issues/9091

Out of curiousity, I tried to drag and drop the saved cells from the viewer onto napari. Unfortunately, I got an error that it can't be read.

You could abstract out the `Load cells` into a napari reader function and this will simplify both the codebase and be useful functionality! https://napari.org/stable/plugins/building_a_plugin/guides.html#reader-example

I think for practical reasons, it would be much more difficult to implement your save function as a writer due to the multiple layers, but if you do #21 it will be much simpler :)

PS its cool that it adds the cell types on top and continues the cycle, I like that functionality, at least.
