### 2025-11-14 pnewstein/napari-3d-counter#15 @pnewstein
I'm having trouble replicating this bug. I'm similarly using
`uvx --with napari-3d-counter 'napari[pyqt6,optional]'` to create an environment and launch napari

Perhaps it's an operating system issue. I'm using arm64 macOS. What are you using? I may be able to find a matching computer for testing.
