### 2025-11-14 pnewstein/napari-3d-counter#15 @TimMonko
Now I'm reliably replicating it. Any time I press a hotkey, the next point is not being added regardless of the hotkey. So if I'm on type 1, press Q to stay on type 1 (or w to go to type 2), the next click doesn't add a point.

I can confirm this doesn't happen when I press the button from the GUI. As long as the button press and hotkey trigger the same callback, it could be a bug on the napari side where the canvas is losing focus after a hotkey press (this sometimes is different behavior on different OSes, which is why its a guess of mine). Can you link me the code path because I'm a bit tired ATM 😬

Here is my environment on a Windows 11 laptop. I'll also test from another Windows machine later
```
napari: 0.6.6
Platform: Windows-11-10.0.26100-SP0
Python: 3.13.3 (main, May 22 2025, 01:58:53) [MSC v.1943 64 bit (AMD64)]
Qt: 6.10.0
PyQt6: 6.10.0
NumPy: 2.3.4
SciPy: 1.16.3
Dask: 2025.11.0
VisPy: 0.15.2
magicgui: 0.10.1
superqt: 0.7.6
in-n-out: 0.2.1
app-model: 0.4.0
psygnal: 0.15.0
npe2: 0.7.9
pydantic: 2.12.4

OpenGL:
- PyOpenGL: 3.1.10
- GL version: 4.6.0 - Build 32.0.101.7026
- MAX_TEXTURE_SIZE: 16384
- GL_MAX_3D_TEXTURE_SIZE: 2048

Screens:
- screen 1: resolution 1920x1200, scale 1.0

Optional:
- numba: 0.62.1
- triangle: 20250106
- napari-plugin-manager: 0.1.8
- bermuda: 0.1.6
- PartSegCore not installed

Experimental Settings:
- Async: False
- Autoswap buffers: False
- Triangulation backend: Fastest available

Settings path:
- C:\Users\Tim\AppData\Local\napari\vm49ii6T3MPA4wvMDGez1_49b041bdd78bc92755ad5db7f9bb5fff2d3b5de5\settings.yaml

Launch command:
- C:\Users\Tim\AppData\Local\uv\cache\archive-v0\vm49ii6T3MPA4wvMDGez1\Scripts\napari

Plugins:
- napari: 0.6.6 (88 contributions)
- napari-3d-counter: 0.5.5 (12 contributions)
- napari-console: 0.1.4 (0 contributions)
- napari-svg: 0.2.1 (2 contributions)
```
