### 2025-11-14 pnewstein/napari-3d-counter#15 @pnewstein
Thanks! I will check this out on a windows pc.

Below is the relevant code:

When a keybinding associated with a cell type is clicked Count3D.change_state_to is called:

https://github.com/pnewstein/napari-3d-counter/blob/c67851f991f8f5eb79aa4b12306150b5eea6fe6b/src/napari_3d_counter/_widget.py#L448-L458

The keybinding is registered here:
https://github.com/pnewstein/napari-3d-counter/blob/c67851f991f8f5eb79aa4b12306150b5eea6fe6b/src/napari_3d_counter/_widget.py#L405-L415

When the canvas is clicked, Napari3dCounter is notified with an events.data callback defined in Count3D.new_pointer_point

https://github.com/pnewstein/napari-3d-counter/blob/c67851f991f8f5eb79aa4b12306150b5eea6fe6b/src/napari_3d_counter/_widget.py#L341-L366

and registered here:
https://github.com/pnewstein/napari-3d-counter/blob/c67851f991f8f5eb79aa4b12306150b5eea6fe6b/src/napari_3d_counter/_widget.py#L265
