### 2025-11-17 pnewstein/napari-3d-counter#15 @TimMonko
Woah, this is genuinely surprising to me. I'm testing this on a different computer that is almost identical in components, and I'm not able to replicate this issue.

```
napari: 0.6.6
Platform: Windows-11-10.0.26100-SP0
Python: 3.13.0 (main, Oct 16 2024, 00:33:24) [MSC v.1929 64 bit (AMD64)]
Qt: 6.10.0
PyQt6: 6.10.0
NumPy: 2.3.5
SciPy: 1.16.3
Dask: 2025.11.0
VisPy: 0.15.2
magicgui: 0.10.1
superqt: 0.7.6
in-n-out: 0.2.1
app-model: 0.4.0
psygnal: 0.15.0
npe2: 0.7.9
pydantic: 2.12.4

OpenGL:
- PyOpenGL: 3.1.10
- GL version: 4.6.0 NVIDIA 560.94
- MAX_TEXTURE_SIZE: 32768
- GL_MAX_3D_TEXTURE_SIZE: 16384

Screens:
- screen 1: resolution 2560x1440, scale 1.0
- screen 2: resolution 1440x2560, scale 1.0

Optional:
- numba: 0.62.1
- triangle: 20250106
- napari-plugin-manager: 0.1.8
- bermuda: 0.1.6
- PartSegCore not installed

Experimental Settings:
- Async: False
- Autoswap buffers: False
- Triangulation backend: Fastest available

Settings path:
- C:\Users\timmo\AppData\Local\napari\de5TXz6qmHukWFTc0paFP_e40503ab4ddbe4533962344aa484c45c7e819a51\settings.yaml

Launch command:
- C:\Users\timmo\AppData\Local\uv\cache\archive-v0\de5TXz6qmHukWFTc0paFP\Scripts\napari

Plugins:
- napari: 0.6.6 (88 contributions)
- napari-3d-counter: 0.5.5 (12 contributions)
- napari-console: 0.1.4 (0 contributions)
- napari-svg: 0.2.1 (2 contributions)
```
