### 2025-11-18 616e500
:feat added examples for each of the widgets
in order to respond cudmore's review:
https://github.com/openjournals/joss-reviews/issues/9091

they are found in the new examples directory
This also deletes the two confusing example scripts on the top level of
the repo
