### 2025-11-18 0b481ee
:fix added installation instructions
