### 2025-11-18 90c3a45
:feat added tooltips to aux plugins
A suggestion from Cudmore

https://github.com/openjournals/joss-reviews/issues/9091
