### 2025-12-02 c7168cb
:fix closes #23
config.face_color was getting its color from
self.layer.currer_border_color. It should be current_face_color

this was creating points with the incorrect apperence
