### 2025-12-03 a141142
:fix colorblind friendly default colors
fixes #18

Uses Paul Tol vibrent colors: https://github.com/Descanonge/tol_colors

also uses the default sequence in the default config (the config which
is used when celltype_config is None)
