### 2025-12-03 e9103ee
:fix use native file dialog where possible
when not in ipython

closes #20
