### 2025-12-03 fb5019b
:fix correct version specification in comment
