### 2025-12-16 e68488f
:feat added a napari reader function
dragging the right csv file will allow you to load the data into
Count3D. The points will be added to the existing Count3D widget if
possible, or else, it will create a new one.

closes #22
issue #24 still stands, but the severity is low
