+++ README.md
@@ -11,14 +11,16 @@ A plugin for manually counting objects in 3D images
 
 ![small](https://github.com/pnewstein/napari-3d-counter/assets/30813691/9d524c31-f23b-4b34-bcb6-ec3bb415cdae)
 
+Documentation can be found below and at the [this site](https://pnewstein.github.io/napari-3d-counter.html)
+
 ----------------------------------
 
-This [napari] plugin was generated with [Cookiecutter] using [@napari]'s [cookiecutter-napari-plugin] template.
+This [napari](https://github.com/napari/napari) plugin was generated with [Cookiecutter](https://github.com/audreyr/cookiecutter) using @napari's [cookiecutter-napari-plugin](https://github.com/napari/cookiecutter-napari-plugin) template.
 
 ## Prerequisites
 
 It is recommended to use conda to install Napari-3D-Counter and napari.
-Installation instructions for the miniconda distribution of conda can be found here: 
+Installation instructions for the miniconda distribution of conda can be found here:
 
 [https://www.anaconda.com/docs/getting-started/miniconda/install](https://www.anaconda.com/docs/getting-started/miniconda/install)
 
@@ -28,7 +30,7 @@ Installation instructions for the miniconda distribution of conda can be found h
 
 
 You can install `napari-3d-counter` via conda
-    
+
     conda create -n n3dc-env -c conda-forge -y napari napari-3d-counter pyqt python=3.12
     conda activate n3dc-env
 
@@ -37,12 +39,28 @@ pip
 
     pip install napari-3d-counter
 
+<details>
+
+<summary>Troubleshooting</summary>
+
+If above conda and pip installs do not work. follow [the
+instructions](https://napari.org/stable/tutorials/fundamentals/installation.html)
+to get a working napari installation. Then install napari-3d counter into that
+environment:
+
+    conda activate napari-env
+    pip install napari-3d-counter
+
+</details>
 
 ##  Count3D Usage
 
 1. First launch napari with the `napari` command.
 1. Count3D can be launched from the plugin menu of napari, or through the
    command palette (Ctrl+Shift+P). Select Count3D.
+
+> [!TIP]
+> Count can also be launched with `examples/launch_count_3d.py`
 
 This will spawn several [Points Layers](https://napari.org/stable/howtos/layers/points.html):
 
@@ -72,9 +90,18 @@ screen) will be incremented.
 
 <summary>Troubleshooting</summary>
 
+#### No cell was added
+
 - Ensure that `Point adder` layer is selected
 - Ensure that `Add points` tool is selected
+- Ensure that there are no image layers obscuring the layers spawned by napari
 - Click on the viewer where you would like the point to be added
+
+
+#### The cell marker is the wrong size
+
+- See [Change appearance of a cell type](https://github.com/pnewstein/napari-3d-counter#change-appearance-of-a-cell-type)
+- Use the `point size` slider
 
 </details>
 
@@ -85,7 +112,10 @@ https://github.com/pnewstein/napari-3d-counter/assets/30813691/745d495e-1d18-43d
 ### Changing cell type
 
 You can change the currently selected cell type by clicking on that cell type's
-button. This change will be reflected in the GUI. Additionally, the keyboard
+button. This change will be reflected in the the color and text of the first box in the GUI. 
+Now any clicks to the canvas will be added to that type.
+
+Additionally, the keyboard
 shortcut for that cell type can be used. Keyboard shortcuts are listed on the
 button, and are "q", "w", "e", "r", "t", "y" by default
 
@@ -96,7 +126,7 @@ https://github.com/pnewstein/napari-3d-counter/assets/30813691/844d04ce-2795-422
[…]
