### 2025-12-17 1fc157a
fix: remove unused makefile rule
