+++ setup.cfg
@@ -24,7 +24,7 @@ classifiers =
     Topic :: Scientific/Engineering :: Image Processing
 project_urls =
     Bug Tracker = https://github.com/pnewstein/napari-3d-counter/issues
-    Documentation = https://github.com/pnewstein/napari-3d-counter#README.md
+    Documentation = http://localhost:8001/napari-3d-counter.html
     Source Code = https://github.com/pnewstein/napari-3d-counter
     User Support = https://github.com/pnewstein/napari-3d-counter/issues
 
