### 2025-12-17 #9091 @pnewstein
@cudmore

Thanks for your review. I found your comments to be constructive, and implementing these changes has greatly improved napari-3d-counter.

Below are my responses to your comments:

______________

> Can you better define what you mean by 'segments' in the following:

I have defined segments in the manuscript

> I am not familiar with Napari 'spots', see "Overall, using Napari-3D-Counter is likely to save the expert annotator’s time over using native Napari spots". I found a napari page on spots but am still a bit confused. Can you clarify this?

Thank you for catching this error. I intended to write napari Points. This has been changed in the manuscript.

> The readme needs a description of how to install from source and run from the cloned codebase. Right now there is only documentation for conda/pip install from PyPi/Conda-Forge.

I have added these instructions to the

    ## Contributing

heading

> For me, the readme install did not work on macOS with an apple silicon (M1) cpu.
>
> Following the readme to install
>
> conda create -n n3dc-env -c conda-forge -y napari napari-3d-counter pyqt
>
>
> On an M1 macOS machine, this does not finish within, idk, 5 minutes. It gets stuck on:
>
> Solving environment: failed with repodata from current_repodata.json, will retry with next repodata source.
> Collecting package metadata (repodata.json): done
> Solving environment: \
>
> I am not sure if your conda create is correct, usually it needs a python= to install a particular version of python?
>
> The following worked for me on an M1 macOS apple-silicon machine. The 'napari[all]' in single quotes is needed when running in zhs versus bash
>
> `# deactivate any existing conda env`
> conda deactivate
>
> `# delete conda env if it exists`
> conda env remove -n n3dc-env -y
>
> `# create and activate a python 3.11 conda env named n3dc-env`
> conda create -y -n n3dc-env -c conda-forge python=3.11
> conda activate n3dc-env
>
> `# install napari and plugin with pip`
> pip install 'napari[all]' napari-3d-counter

On my system, (also arm64 macos) the solver is able to pick an appropriate python version. However, I have added a python version specification to try to reduce failed installs.

I have added a troubleshooting dropdown in the Installation section section with advice on installation, also linking to napari's troubleshooting guide as I expect my package to install fine on any environment with a recent napari.

> 3. I see two sample scripts including gui_demo.py, and launch_napari.py.
>
>     There is no mention of what these scripts do in the readme, and they contaminate the root of the GitHub repo.
>     I would like to see a description in the readme telling the user what these scripts do.
>     I would also like these scipts moved into a folder such as, idk, examples/.
>
> I see these scripts generate synthetic data. It would be best to provide the user with a more explicit sample images/volumes. You could make a 3d 'blob' numpy array and save it as a tif file. You could include this tif file in the repo or add to the repo or readme the code to create it. My point is, your scripts mix multiple functions together including: (i) making a sample image, and (ii) runing the Napari GUI. These functions should be functionaly/programatically separated.

I have added an examples directory with one script per widget. This is modeled after napari's own examples directory

> 4.
>
> I would like to see more description and user recipes for the 'Auxiliary plugins'. These should be better described in the readme.
>
> Each gui element in these auxiliary plugins should have a tooltip for these gui elements to help the user understand what the purpose is.

I have added the tool-tips. The user recipes demonstrating the purposes of each of the plugins are now in the `examples` directory

> Ingress Points: From the manuscript:
>
>     Takes a native napari points layer, perhaps created by automated labeling, and turns them into a counted cell type in Count3D". I was able to get this to work by creating my own napari points layer, adding some points and then running ingress points on that layer.
>
> As far as I understand, this takes an existing points layer created by the user and brings it into one of the N3DC layers. I would like to see more of a recipe for using ingress points. This could be added to the readme.

 I have added `examples/ingress_points.py`, which illustrates functionality

>
> Reconstruct Selected: From the manuscript:
>
>     Can be used to aid in visualizing cells: if a user has a Label layer containing some labels corresponding to cells of interest, and other labels not corresponding, perhaps created by automated segmentation, ReconstructSelected will convert labels containing a Count3D cell into an image layer, which can then be used to create a 2D or 3D image
>
> I was not able to use the Reconstruct Selected? I did not have a labeled layer handy. Like the other auxiliary plugins, can you provide a recipe with more detail in the readme? In this case it might be good to have code to create a simple example of a labeled layer/image and how to use it in this plugin.

 I have added `examples/reconstruct_selected.py`, which sets up the widget and illustrates functionality

> Split On Shapes: From the manuscript:
>
>     Splits labels of cell types based on spatial information. For example, if a user wants to quantify the distribution of cells of multiple types across segments, they can use a napari Shapes layer to define all the segments in the X and Y axes, and SplitOnShapes will return a count of each cell type within each shape.
>
> As with the other auxiliary plugins, I would like to see more user documentation so the user can use this auxiliary plugin.

I have added `examples/split_on_shapes.py`, which illistruates functionality

> 5.
> In adition to a local git clone install in the readme. You need to include a short recipe on how to (i) run unit tests, (ii) how to contribute via a pull request, and (iii) not a recipe but mention community standards.

This has been added to the contributing section

> 6. Maybe add a troubleshooting section
>
> A troubleshooting section would be a good place to throw any small gui glitches you come acress. At least until they are fixed in the workflow/gui of the plugin.
>
>     For example, if the plugin is run, then an image is opened, the image is on top of the plugins points layer. User needs to rearrange the napari layers such that the plugin points layer is on top and visible.
>

 I have added troubleshooting dropdowns to the Installation and `Adding a Cell` headers. This includes your example as well as a few other problems I have noticed people encountering

> 7.
>
> The readme contains movies of how to use the GUI, this is excellent. I mentioned this earlier but I would like to see more descriptive recipes of the steps involved in using the plugin. These should step through what we are looking at in the screenshot movies.

I have added more detail to the readme stepping through all of the steps depicted in the videos, as well as those examples which give more detail.

> 8.
>
> Although not needed for this revision. I would strongly suggest making a dedicated web based static site for documentation. The readme gets long with all the details of using the plugin. You could easily implement this in MkDocs, include it in the repo and have GitHub workflows or even local scripts that build and publish the documentation to GitHub gh-pages.

I have uploaded a standalone site to https://pnewstein.github.io/napari-3d-counter.html which gives the readme a nice table of contents, making it easier to find what you're looking for.
