### 2025-12-17 pnewstein/napari-3d-counter#15 @pnewstein
I have also not been able to replicate this issue on the Windows PC I have access to
