### 2026-01-06 pnewstein/napari-3d-counter#17 @TimMonko
Great! This works now and make sense. :)
