### 2026-01-07 pnewstein/napari-3d-counter#25 @MaximLippeveld
Add a statement of need to the documentation
Issue with regards to the [JOSS manuscript review](https://github.com/openjournals/joss-reviews/issues/9091)

I think you can just re-use the statement of need from the manuscript.
