### 2026-01-07 pnewstein/napari-3d-counter#26 @MaximLippeveld
Improve overall explanation of ReconstructSelected plugin
Issue with regards to the [JOSS manuscript review](https://github.com/openjournals/joss-reviews/issues/9091)

The explanation is not very clear to me. Maybe use some concrete example to explain it?

I suggest to update the documentation as well as the popup labels in the napari plugin itself.
