### 2026-01-07 pnewstein/napari-3d-counter#27 @MaximLippeveld
Remove subjective sentence on plugin usage from the manuscript
Issue with regards to the [JOSS manuscript review](https://github.com/openjournals/joss-reviews/issues/9091)

The manuscript contains the following sentence: "This widespread use is likely due to the plugin's ergonomics and usability." While I believe this could be true, I find that the manuscript should only contain the proven facts (such as the citations and number of downloads that you mention). Therefore, I would remove this sentence.
