### 2026-01-07 #9091 @MaximLippeveld
@pnewstein Great work with the plugin and manuscript! I think you made a valuable contribution the the napari ecosystem. I have created a few minor issues on the plugin repo.
