### 2026-01-07 73b3584
:fix removed subjective sentence
closes #27
