### 2026-01-12 21a89f0
:fix improved explanation of ReconstructSelected
closes #26
