### 2026-01-12 4d5b843
:fix added statement of need to docs
closes #25
