### 2026-01-23 60cee89
:fix use scipy Delaunay to compute hulls
bugs in shapes layers makes it so that using the internal
Shapes.get_value fails. Its straightforward to calculate these hulls
with scipy in these cases. But I should find the source of the napari
bug
