### 2026-01-23 ed9ba62
:fix uncheck ingress points after click
