### 2026-01-23 3fcf926
:fix add cell_type column to split_on_shapes
it is strange to use the index for this
