### 2026-01-29 d98449a
:fix bugfix to use new segmentation
correct fstring
also work on final_assignment 0
correct `all` comparison
