### 2026-01-29 b965d9b
:fix allow differend column orders
this allows the plugin to read more datatypes
it uses set logic to compare a manualy split first line to the neccisary
columns
