### 2026-03-03 #9091 @pnewstein
@adamltyson Do you know what went wrong here?

The old links to the draft manuscripts are also now broken: https://github.com/openjournals/joss-reviews/issues/9091#issuecomment-3667521029

It seems as though the 09091 branch corresponding to this review on the https://github.com/openjournals/joss-papers repo has been deleted.

How might we proceed?
