### 2026-03-04 #9091 @arfon
> Sorry @pnewstein I have no idea! I will ask the JOSS dev team.

Looks like it was a transient issue with the GitHub API.
