### 2026-03-05 #9091 @pnewstein
@adamltyson Sorry about that! The title on the zenodo page has been corrected to: Napari-3D-Counter: A manual cell counter for napari
