### 2026-03-15 #9091 @MaximLippeveld
> [@pnewstein](https://github.com/pnewstein) congratulations on this JOSS publication!
>
> Thank you [@adamltyson](https://github.com/adamltyson) for editing!
>
> And a special thank you to the reviewers [@TimMonko](https://github.com/TimMonko), [@MaximLippeveld](https://github.com/MaximLippeveld) !!

Congratulations @pnewstein! 🥳

Happy to help!
