### 2026-03-23 #9091 @pnewstein
Thanks all for your reviews! They have really helped improve napari-3d-counter.

And thanks adamltyson for guiding the process, as well as everyone in JOSS for maintaining this important resource!
